requirements.txt
tensorflow
numpy
pillow
discord.py

